test-with-pom - move user ui tests into user dir
