+++ e2e/user/user-can-cancel-add-dog.spec.ts
@@ -1,8 +1,8 @@
 import { test, expect } from "@playwright/test";
-import { registerTestUser } from "./_lib/sequences/register-test-user";
-import { getTestBirthday } from "./_lib/e2e-test-utils";
-import { generateTestDogName } from "./_lib/e2e-test-utils";
-import { urlOf } from "./_lib/e2e-test-utils";
+import { registerTestUser } from "../_lib/sequences/register-test-user";
+import { getTestBirthday } from "../_lib/e2e-test-utils";
+import { generateTestDogName } from "../_lib/e2e-test-utils";
+import { urlOf } from "../_lib/e2e-test-utils";
 import { RoutePath } from "@/lib/route-path";
 
 test("user can register, login, add dog but cancel, and not should not see new dog in my-pets", async ({
